refactor(20335): Renomeia arquivo modelo membro_associacao
O arquivo estava no plural (membros_associacao) e o padrão é estar no singular.
