+++ sme_ptrf_apps/core/models/__init__.py
@@ -6,12 +6,12 @@ from .ata import Ata
 from .conta_associacao import ContaAssociacao
 from .demonstrativo_financeiro import DemonstrativoFinanceiro
 from .fechamento_periodo import STATUS_ABERTO, STATUS_FECHADO, STATUS_IMPLANTACAO, FechamentoPeriodo
-from .membros_associacao import MembroAssociacao
+from .membro_associacao import MembroAssociacao
 from .observacao import Observacao
 from .parametros import Parametros
 from .periodo import Periodo
 from .prestacao_conta import STATUS_ABERTO, STATUS_FECHADO, PrestacaoConta
 from .relacao_bens import RelacaoBens
+from .tag import Tag
 from .tipo_conta import TipoConta
 from .unidade import Unidade
-from .tag import Tag
